+++ server/db/db.js
@@ -146,7 +146,38 @@ async function getClients(pgClient) {
       return null
     }
 
-    return await Promise.all(results.rows.map(r => createClientFromRow(r, pgClient)))
+    return await Promise.all(results.rows.map(r => createClientFromRow(r)))
+  } catch (err) {
+    helpers.log(err.toString())
+  }
+}
+
+async function getActiveClients(pgClient) {
+  try {
+    const results = await helpers.runQuery(
+      'getActiveClients',
+      `
+      SELECT c.*
+      FROM clients c
+      INNER JOIN (
+        SELECT DISTINCT client_id AS id
+        FROM locations
+        WHERE is_sending_alerts AND is_sending_vitals
+      ) AS l
+      ON c.id = l.id
+      WHERE c.is_sending_alerts AND c.is_sending_vitals
+      ORDER BY c.display_name;
+      `,
+      [],
+      pool,
+      pgClient,
+    )
+
+    if (results === undefined) {
+      return null
+    }
+
+    return await Promise.all(results.rows.map(r => createClientFromRow(r)))
   } catch (err) {
     helpers.log(err.toString())
   }
@@ -898,7 +929,7 @@ async function updateClient(
 
     helpers.log(`Client '${displayName}' successfully updated`)
 
-    return await createClientFromRow(results.rows[0], pgClient)
+    return await createClientFromRow(results.rows[0])
   } catch (err) {
     helpers.log(err.toString())
   }
@@ -1465,6 +1496,7 @@ module.exports = {
   getAllSessionsFromLocation,
   getClientWithClientId,
   getClients,
+  getActiveClients,
   getCurrentTime,
   getDataForExport,
   getHistoricAlertsByAlertApiKey,